Extract tests out to separate files
