+++ medicines/pars-upload/cypress/integration/new-par.js
@@ -1,0 +1,163 @@
+import {
+  addAndDeleteSubstances,
+  addAndDeleteProducts,
+  completeUploadForm,
+  completeUploadFile,
+  mockSuccessfulSubmission,
+  setUp,
+} from '../support/shared'
+
+/* eslint-env node, mocha */
+
+/* global Cypress, cy */
+
+setUp()
+
+const parsUrl = Cypress.env('PARS_UPLOAD_URL')
+const baseUrl = Cypress.config().baseUrl
+
+describe('PARs upload', () => {
+  it('can add and delete multiple substances', () => {
+    cy.visit('/new-par')
+    let uploadData = {
+      brand: 'Ibuprofen pills',
+      strength: 'Really powerful stuff',
+      doseForm: 'some form',
+      substance1: 'Ibuprofen',
+      substance2: 'Paracetamol',
+      substance3: 'Temazepam',
+    }
+    addAndDeleteSubstances(uploadData)
+  })
+
+  it('can add and delete multiple products', () => {
+    cy.visit('/new-par')
+    let uploadData = {
+      brand: 'Ibuprofen pills',
+      strength: 'Really powerful stuff',
+      doseForm: 'some form',
+      substance1: 'Ibuprofen',
+      substance2: 'Paracetamol',
+      licence: { type: 'THR', part_one: '12345', part_two: '6789' },
+    }
+    addAndDeleteProducts(uploadData)
+  })
+
+  it('review page shows the correct information', () => {
+    cy.visit('/new-par')
+    let uploadData = {
+      brand: 'Ibuprofen pills',
+      strength: 'Really powerful stuff',
+      doseForm: 'some form',
+      substances: ['Ibuprofen', 'Paracetamol'],
+      licence: { type: 'THR', part_one: '12345', part_two: '6789' },
+    }
+    completeUploadForm(uploadData)
+
+    const fileName = 'rabbit-anti-human-stuff.pdf'
+    completeUploadFile(fileName)
+
+    cy.findAllByText('Check your answers before sending the report')
+      .not('title')
+      .should('have.length', 1)
+
+    cy.findByText('Brand/Generic name')
+      .parent()
+      .within(() => {
+        cy.findByText(uploadData.brand).should('exist')
+      })
+
+    cy.findByText('Strength')
+      .parent()
+      .within(() => {
+        cy.findByText(uploadData.strength).should('exist')
+      })
+
+    cy.findByText('Pharmaceutical dose form')
+      .parent()
+      .within(() => {
+        cy.findByText(uploadData.doseForm).should('exist')
+      })
+
+    cy.findByText('Active substances')
+      .parent()
+      .within(() => {
+        cy.findByText(uploadData.substances.join(', ')).should('exist')
+      })
+
+    cy.findByText('Licence number')
+      .parent()
+      .within(() => {
+        cy.findByText(
+          `${uploadData.licence.type} ${uploadData.licence.part_one}/${uploadData.licence.part_two}`
+        ).should('exist')
+      })
+
+    cy.findByText('Document')
+      .parent()
+      .parent()
+      .within(() => {
+        cy.findByText('Document name')
+          .parent()
+          .within(() => {
+            cy.findByText(fileName).should('exist')
+          })
+      })
+
+    const licence_str = `${uploadData.licence.type} ${uploadData.licence.part_one}/${uploadData.licence.part_two}`
+    const product_title = `${uploadData.brand}, ${uploadData.strength}, ${uploadData.doseForm}, ${licence_str}`.toUpperCase()
+
+    cy.findByText(product_title)
+      .parent()
+      .within(() => {
+        cy.findAllByText('Change').last().click()
+      })
+
+    cy.findAllByText('New Public Assessment Report')
+      .not('title')
+      .should('have.length', 1)
+
+    cy.findByLabelText('Brand/Generic name').should(
+      'have.value',
+      uploadData.brand
+    )
+  })
+
+  it('can submit the form sucessfully', () => {
+    if (parsUrl) {
+      cy.log('Mocking form submissions endpoint')
+
+      cy.server()
+
+      mockSuccessfulSubmission(baseUrl, parsUrl)
+    }
+
+    cy.visit('/new-par')
+
+    let uploadData = {
+      brand: 'Ibuprofen pills',
+      strength: 'Really powerful stuff',
+      doseForm: 'some form',
+      substances: ['Ibuprofen', 'Paracetamol'],
+      licence: { type: 'THR', part_one: '12345', part_two: '6789' },
+    }
+    completeUploadForm(uploadData)
+
+    const fileName = 'rabbit-anti-human-stuff.pdf'
+    completeUploadFile(fileName)
+
+    cy.findAllByText('Check your answers before sending the report')
+      .not('title')
+      .should('have.length', 1)
+
+    cy.findByText('Accept and send').click()
+
+    cy.findAllByText('Submission complete')
+      .not('title')
+      .should('have.length', 1)
+
+    cy.findByText('Submit another report').click()
+
+    cy.findByText('What are you doing today?').should('exist')
+  })
+})
